Move components in the events/ subfolders into appropriate subfolders Move all container components into a container/ subfolder Move all presentational components into a presentational/ subfolder Modify paths to components in other components that import them
