+++ client/components/AdminProfile.jsx
@@ -7,7 +7,7 @@ import * as utilityActions from '../actions/utilityActions';
 import * as eventActions from '../actions/eventActions';
 import * as centerActions from '../actions/centerActions';
 import Search from './Search';
-import EventsListWithImage from './events/EventsListWithImage';
+import EventsListWithImage from './events/presentational/EventsListWithImage';
 import CenterList from './centers/presentational/CenterList';
 import Preloader from './common/Preloader';
