+++ client/components/UserProfile.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import Search from './Search';
-import EventsListWithImage from './events/EventsListWithImage';
+import EventsListWithImage from './events/presentational/EventsListWithImage';
 import Preloader from './common/Preloader';
 import * as eventActions from '../actions/eventActions';
