+++ client/components/events/container/SingleEvent.jsx
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import * as eventActions from '../../actions/eventActions';
-import EventDetails from './EventDetails';
+import * as eventActions from '../../../actions/eventActions';
+import EventDetails from '../presentational/EventDetails';
 
 class SingleEvent extends Component {
   componentDidMount() {
